refactor(setLocale): Rename file for clarity and export default setLocale
BREAKING CHANGE: customLocale.js is now setLocale.js
