+++ test/setLocale.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-template-curly-in-string */
-import setLocale from '../src/customLocale';
+import { setLocale } from '../src';
 
 describe('Custom locale', () => {
   it('should get default locale', () => {
